requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2.1
marshmallow==3.19.0
packaging==23.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
